refactor(upstream): split aci/upstream.rs into focused submodules
Break the 2021-line src/aci/upstream.rs into a directory module:

- mod.rs       — core request/response types, UpstreamBackend trait, re-exports
- router.rs    — ModelRoute + ModelRouterBackend (+ inline tests)
- openai.rs    — OpenAICompatibleBackend + model-id helpers
- tls.rs       — SPKI-pinning reqwest client + header extraction
- chutes/mod.rs     — ChutesProviderBackend: discovery + E2EE transport
- chutes/session.rs — per-upstream session store (chute-id + nonce pools)
- chutes/crypto.rs  — ML-KEM / ChaCha20-Poly1305 / gzip + decrypting stream

Pure relocation: no logic changes. Public paths preserved via re-export;
cross-module helpers marked pub(super). cargo test --all-targets:
220 passed, 0 failed (unchanged from baseline).
